T64: the port is read before an unsigned short can eat it
parse_host_port answers in an int and the sync loop took an unsigned short, so
70000 arrived as 4464 with nothing said about it, and the ceiling is 63535
rather than 65535 because libtorrent listens two thousand above this port
(V12) and that sum wrapped just as quietly - two nodes on one data plane port,
silently. V12 was simply false above 63535.

check_listen_address reads the port itself now: absent, non numeric, or
outside 1-63535 is a critical line and EXIT_FAILURE. The narrowing at the
sync_loop call is explicit and says why it is safe.

Utils.V68APortOutOfRangeIsRefused with eight spellings and its positive twin,
plus test_v68_port_out_of_range_exits_failure over four. Both red first with
the port half stripped back out: 70000, 65000, 0 and a missing port all ran a
daemon.

V68, V12, V9

+++ src/syncfs.cpp
@@ -736,7 +736,11 @@ auto main(int argc, char *argv[]) -> int try {
 
   auto [host, port] = utils::parse_host_port(args[2]);
 
-  sync_loop(std::move(client), std::move(listener), std::move(host), port);
+  // Narrowed on purpose and only because the check above has already put the
+  // port inside a range that survives it, libtorrent's two thousand more
+  // included (V68).
+  sync_loop(std::move(client), std::move(listener), std::move(host),
+            static_cast<unsigned short>(port));
 
   return EXIT_SUCCESS;
 } catch (zmq::error_t &e) {
